requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
dj-database-url==1.2.0
Django==4.1.4
django-bootstrap4==22.3
django-ckeditor==6.5.1
django-js-asset==2.0.0
gunicorn==20.1.0
notify2==0.3.1
Pillow==9.3.0
psycopg2-binary==2.9.5
pypiwin32==223
python-decouple==3.7
pytube==12.1.0
pywin32==305
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.7
win10toast==0.9
